eliminate readconfig in favor of execution options
