+++ crates/sparrow-interfaces/src/lib.rs
@@ -7,8 +7,10 @@
     clippy::print_stderr
 )]
 
+mod execution_options;
 mod source;
 mod source_error;
 
+pub use execution_options::*;
 pub use source::*;
 pub use source_error::SourceError;
