+++ crates/sparrow-interfaces/src/source.rs
@@ -4,7 +4,7 @@ use arrow_schema::{DataType, SchemaRef};
 use futures::stream::BoxStream;
 use sparrow_batch::Batch;
 
-use crate::SourceError;
+use crate::{ExecutionOptions, SourceError};
 
 /// Trait implemented by sources.
 pub trait Source: Send + Sync {
@@ -19,11 +19,11 @@ pub trait Source: Send + Sync {
     /// Parameters:
     /// * `projected_datatype`: The datatype of the data to produce. Note that
     ///    this may differ from the prepared type.
-    /// * `read_config`: Configuration for the read.
+    /// * `execution_options`: Options for the entire execution.
     fn read(
         &self,
         projected_datatype: &DataType,
-        read_config: Arc<ReadConfig>,
+        execution_options: Arc<ExecutionOptions>,
     ) -> BoxStream<'static, error_stack::Result<Batch, SourceError>>;
 
     /// Allow downcasting the source.
@@ -47,23 +47,4 @@ impl SourceExt for &Arc<dyn Source> {
     fn downcast_source_opt<T: Source + 'static>(&self) -> Option<&T> {
         self.as_any().downcast_ref()
     }
-}
-
-/// Defines the configuration for a read from a source.
-#[derive(Clone, Debug)]
-pub struct ReadConfig {
-    /// If true, the read will act as an unbounded source and continue reading
-    /// as new data is added. It is on the consumer to close the channel.
-    ///
-    /// If false, the read will act as a bounded source, and stop once the set
-    /// of data available at the time of the read has been processed.
-    pub keep_open: bool,
-    /// Optional timestamp in nanos at which to start reading.
-    ///
-    /// Defaults to the earliest available timestamp.
-    pub start_time: Option<i64>,
-    /// Optional timestamp in nanos at which to end reading.
-    ///
-    /// Defaults to reading until the source is closed.
-    pub end_time: Option<i64>,
 }
